refactor(inflight): extract store-side CAS clusters, retire giant (#3835)
Verbatim MOVE of the residual store-side compare-and-set clusters out of the
`inflight.rs` module root into two capped children plus the existing store
primitive layer, retiring inflight.rs from the giant-file registry.

Extraction (zero behavior change, pure move + visibility/path adjustment):
- inflight/save_store.rs (new, 459 prod LoC): the CAS "save" cluster —
  save_inflight_state{,_create_new,_if_absent,_if_matches_identity},
  the rebind-adoption variants, GuardedSaveOutcome, CreateNewInflightError,
  and their _in_root explicit-root seams.
- inflight/clear_store.rs (new, 849 prod LoC): the CAS "clear"/abandon cluster —
  clear_inflight_state{,_if_matches,_if_matches_identity,_after_delivery,
  _zero_owned}, clear_rebind_origin_*, refresh_inflight_last_offset_*,
  request_inflight_abandon_*, enqueue_abandon_request_for_row,
  normalize_response_sent_offset, GuardedClearOutcome, and their _in_root seams.
- inflight/store.rs (+242): the shared persist/lock + save-side validation
  primitives (validate_inflight_state_for_save, load_inflight_state_unlocked,
  persist_under_lock{,_inner,_preserving_updated_at}), joining the existing
  flock/path layer so the CAS children and sibling modules consume one layer.

Root inflight.rs: 2339 -> 914 prod LoC (< 1000) — no longer a prod-giant.

Facade re-exports mirror each moved symbol at its ORIGINAL visibility (3-tier:
pub(crate) / pub(in crate::services::discord) / pub(super)), so every external
flat `inflight::SYMBOL` path stays byte-identical. `pub(super)` seams are
declared `pub(in crate::services::discord)` in the children (the absolute
spelling of the parent's original `pub(super)`) so re-exporting them does not
widen the surface; test-only `_in_root` seams keep `pub(super)` + `#[cfg(test)]`
re-import. Public API is not widened.

The #3016 core hot files (turn_bridge/mod.rs, tmux_watcher.rs,
session_relay_sink.rs, turn_finalizer.rs, tui_prompt_relay.rs) are byte-
untouched — they use only flat `inflight::` paths, resolved by the re-exports.
The offset-monotonic invariant recorder/severity gate stays in the module root
(shared by save + clear); no function body is altered.

Giant retirement: removed from giant_file_registry.toml (frozen list +
[[entry]]), audit_maintainability_giant_baseline.toml, and the
change-surfaces.md freeze entry; regenerated inventory docs.

Gates: cargo fmt --check clean; cargo check --lib clean (0 new warnings, warning
set identical to origin/main); inflight + high_risk_recovery tests match
origin/main exactly; test_inventory_giant_split (34) OK; audit_maintainability
--check OK; hotfile_ratchet OK (core ceilings unchanged).

Closes #3835
