+++ src/services/discord/inflight.rs
@@ -95,6 +95,80 @@ pub(in crate::services::discord) use self::orphan_relay_reclaim::{
     mark_session_bound_relay_delivered_locked, session_bound_relay_external_input_orphan_shape,
 };
 
+// #3835: the store-side CAS "save" and "clear" clusters moved into two capped
+// children so this hot state parent retires from the giant-file registry. Every
+// symbol is re-exported at its ORIGINAL visibility below (3-tier: pub(crate) /
+// pub(in ...discord) / pub(super)), so external flat `inflight::SYMBOL` paths
+// (incl. the #3016 core hot files) stay byte-identical. Public API is not
+// widened. The shared persist/validate primitives moved to `store.rs`.
+mod clear_store;
+mod save_store;
+
+// Shared persist/validate primitives (moved to `store.rs`): re-imported at the
+// parent's original inflight-private visibility so sibling children resolve
+// `super::persist_under_lock` / `super::load_inflight_state_unlocked` /
+// `super::persist_under_lock_preserving_updated_at` and the CAS children resolve
+// `validate_inflight_state_for_save` (via `use super::*`) unchanged.
+use self::store::{
+    load_inflight_state_unlocked, persist_under_lock, persist_under_lock_preserving_updated_at,
+    validate_inflight_state_for_save,
+};
+
+// Save cluster re-exports (original visibility mirrored). The save child declares
+// these `pub(in crate::services::discord)` (the absolute spelling of the parent's
+// original `pub(super)`), so this `pub(super)` re-export does not widen the surface.
+pub(super) use self::save_store::{
+    CreateNewInflightError, save_inflight_state, save_inflight_state_create_new,
+    save_inflight_state_if_absent,
+};
+pub(in crate::services::discord) use self::save_store::{
+    GuardedSaveOutcome, save_existing_inflight_rebind_adoption_if_matches_identity,
+    save_existing_inflight_rebind_adoption_with_offset_rebase_if_matches_identity,
+    save_inflight_state_if_matches_identity,
+};
+// Explicit-root save seams reached only by the parent's / siblings' test modules.
+#[cfg(test)]
+use self::save_store::{
+    save_existing_inflight_rebind_adoption_if_matches_identity_in_root,
+    save_existing_inflight_rebind_adoption_with_offset_rebase_if_matches_identity_in_root,
+    save_inflight_state_if_matches_identity_in_root, save_inflight_state_in_root,
+};
+
+// Clear/Abandon cluster re-exports (original visibility mirrored).
+pub(crate) use self::clear_store::{
+    GuardedClearOutcome, clear_inflight_state, clear_inflight_state_if_matches,
+    clear_inflight_state_if_matches_zero_owned, request_inflight_abandon_if_matches,
+    request_inflight_abandon_if_matches_zero_owned,
+};
+pub(in crate::services::discord) use self::clear_store::{
+    clear_inflight_state_if_matches_identity,
+    clear_inflight_state_if_matches_identity_after_delivery,
+    clear_rebind_origin_inflight_state_if_matches_identity,
+    refresh_inflight_last_offset_if_matches_identity,
+};
+// `clear_*_in_root` seams reached by inflight-core in production (health recovery
+// engine): the clear child declares them `pub(in crate::services::discord)`.
+pub(super) use self::clear_store::{
+    clear_inflight_state_if_matches_in_root, clear_inflight_state_if_matches_zero_owned_in_root,
+};
+// `normalize_response_sent_offset` is consumed by the `watcher_state` sibling in
+// production; `super::abandon_request_store` is a discord-level module the moved
+// abandon-enqueue path references — both re-imported (non-test) so the moved code
+// resolves through `super::*` / `super::abandon_request_store` unchanged.
+use self::clear_store::normalize_response_sent_offset;
+use super::abandon_request_store;
+// Explicit-root clear seams reached only by the parent's / siblings' test modules.
+#[cfg(test)]
+use self::clear_store::{
+    clear_inflight_state_if_matches_identity_after_delivery_in_root,
+    clear_inflight_state_if_matches_identity_in_root,
+    clear_inflight_state_if_matches_tmux_response_in_root,
+    clear_rebind_origin_inflight_state_if_matches_identity_in_root,
+    refresh_inflight_last_offset_if_matches_identity_in_root,
+    request_inflight_abandon_if_matches_in_root,
+    request_inflight_abandon_if_matches_zero_owned_in_root, row_has_finalizable_placeholder,
+};
+
 use finalizer_identity::{
     backfill_finalizer_turn_id_under_lock, parse_inflight_state_content,
     parse_inflight_state_content_with_finalizer_backfill, read_inflight_state_content,
@@ -284,1505 +358,6 @@ fn record_inflight_invariant_with_severity(
         },
         severity,
     )
-}
-
-fn validate_inflight_state_for_save(
-    root: &Path,
-    path: &Path,
-    state: &InflightTurnState,
-    code_location: &'static str,
-) -> bool {
-    let offset_in_bounds = state.response_sent_offset <= state.full_response.len()
-        && state
-            .full_response
-            .is_char_boundary(state.response_sent_offset);
-    record_inflight_invariant(
-        offset_in_bounds,
-        state,
-        "response_sent_offset_in_bounds",
-        code_location,
-        "inflight response_sent_offset must stay within full_response",
-        serde_json::json!({
-            "response_sent_offset": state.response_sent_offset,
-            "full_response_len": state.full_response.len(),
-            "path": path.display().to_string(),
-        }),
-    );
-    debug_assert!(
-        offset_in_bounds,
-        "inflight response_sent_offset must stay within full_response"
-    );
-
-    let Ok(existing_content) = fs::read_to_string(path) else {
-        return true;
-    };
-    let Ok(existing) = serde_json::from_str::<InflightTurnState>(&existing_content) else {
-        return true;
-    };
-
-    // #3154 — OBSERVE-ONLY on the bridge/watcher save path. A legit fresh-turn
-    // reset (different user_msg_id or turn_start_offset) resets
-    // response_sent_offset to 0 on purpose (see InflightTurnState::new), so the
-    // check is gated by SAME turn identity; only a backward move within the same
-    // turn is a violation. We do not skip the write here (that would drop a
-    // legit fresh turn); this mirrors the last_offset_monotonic precedent below.
-    let same_turn_identity = existing.user_msg_id == state.user_msg_id
-        && existing.turn_start_offset == state.turn_start_offset;
-    let monotonic_offset =
-        !same_turn_identity || state.response_sent_offset >= existing.response_sent_offset;
-    // I6 (last_offset_monotonic) — OBSERVE-ONLY on the bridge/watcher save
-    // path. A legit fresh-turn reset (different user_msg_id or
-    // turn_start_offset) lowers last_offset on purpose, so the check is gated
-    // by SAME turn identity; only a backward move within the same turn is a
-    // violation. We do not skip the write here (that would drop a legit fresh
-    // turn); the enforcing variant lives in the standby/refresh path.
-    let last_offset_monotonic = !same_turn_identity || state.last_offset >= existing.last_offset;
-
-    // #3552: when the #3416 enforce guard (below) will SKIP this backward write
-    // and preserve the offset (zero data loss), the offset-monotonic violation
-    // has already been safely handled — record it at WARN instead of ERROR so
-    // the paired `#3416 enforce` WARN is the only operator-facing log, killing
-    // the duplicate ERROR-log noise. When enforce is OFF a GENUINE (non-reset)
-    // backward write actually persists below, so that violation stays ERROR (a
-    // real breach); the legitimate re-stream reset (#3933) is handled separately
-    // just before the records below. Computed BEFORE the records so the severity
-    // is correct; the enforce branch itself (skip + return false) is unchanged.
-    // #3933: a legitimate Gemini/Qwen `RetryBoundary` reset rewinds the SAME
-    // turn's frontier to the start — `full_response` cleared and
-    // `response_sent_offset` back to 0 — to re-stream the answer
-    // (turn_bridge/retry_state.rs::clear_response_delivery_state). That backward
-    // move is NOT a stale-snapshot regression, so the enforce guard must permit
-    // it (the release runs AGENTDESK_DELIVERY_RECORD_AUTHORITY=1; blocking it
-    // drops the re-streamed body). A genuine backward regression carries a
-    // non-empty body, so it never matches this rewind signature and stays
-    // blocked. The signal is derived here from the incoming state — no call-site
-    // change — so the guard stays self-contained.
-    let is_legitimate_full_reset =
-        same_turn_identity && state.full_response.is_empty() && state.response_sent_offset == 0;
-    use crate::services::discord::outbound::delivery_record as dr;
-    let authority = dr::delivery_record_authority_enabled();
-    let enforce_skips_backward_write = dr::authority_blocks_backward_inflight_write(
-        authority,
-        monotonic_offset,
-        last_offset_monotonic,
-        is_legitimate_full_reset,
-    );
-    // #3933: a legitimate full reset PERSISTS its backward write (it is a permitted
-    // re-stream rewind, so the enforce guard does NOT skip it —
-    // `enforce_skips_backward_write` is false). That rewind is intended, not a
-    // data-loss regression, so it must not surface an operator-facing ERROR: treat
-    // it as "safely handled" (WARN) exactly like the enforce-skip case. This is a
-    // severity-label change ONLY — the enforce guard, the debug tripwire (which
-    // still keys off `enforce_skips_backward_write`), and the on-disk schema are
-    // all unchanged.
-    let monotonic_violation_safely_handled =
-        enforce_skips_backward_write || is_legitimate_full_reset;
-    let offset_monotonic_severity =
-        offset_monotonic_invariant_severity(monotonic_violation_safely_handled);
-
-    record_inflight_invariant_with_severity(
-        monotonic_offset,
-        state,
-        "response_sent_offset_monotonic",
-        code_location,
-        "inflight response_sent_offset must not move backwards for the same turn identity",
-        serde_json::json!({
-            "previous": existing.response_sent_offset,
-            "next": state.response_sent_offset,
-            "same_turn_identity": same_turn_identity,
-            "path": path.display().to_string(),
-        }),
-        offset_monotonic_severity,
-    );
-    // #3933: when the enforce guard is about to SKIP this backward write it never
-    // persists, so the debug tripwire has nothing to catch — asserting there would
-    // panic on a write we already discard. Relax the tripwire for that skipped
-    // case only; a backward move that actually PERSISTS (enforce OFF, or a
-    // permitted legitimate reset in release) still trips it, preserving the
-    // tripwire's purpose and every existing observe-only test verbatim.
-    debug_assert!(
-        monotonic_offset || enforce_skips_backward_write,
-        "inflight response_sent_offset must not move backwards for the same turn identity"
-    );
-
-    record_inflight_invariant_with_severity(
-        last_offset_monotonic,
-        state,
-        "last_offset_monotonic",
-        code_location,
-        "inflight last_offset must not move backwards for the same turn identity",
-        serde_json::json!({
-            "previous": existing.last_offset,
-            "next": state.last_offset,
-            "same_turn_identity": same_turn_identity,
-            "path": path.display().to_string(),
-        }),
-        offset_monotonic_severity,
-    );
-    debug_assert!(
-        last_offset_monotonic || enforce_skips_backward_write,
-        "inflight last_offset must not move backwards for the same turn identity"
-    );
-
-    let same_tmux_owner = existing.tmux_session_name.is_none()
-        || state.tmux_session_name.is_none()
-        || existing.tmux_session_name == state.tmux_session_name;
-    record_inflight_invariant(
-        same_tmux_owner,
-        state,
-        "inflight_tmux_one_to_one",
-        code_location,
-        "inflight state for a channel must not drift between tmux sessions",
-        serde_json::json!({
-            "previous_tmux_session_name": existing.tmux_session_name.as_deref(),
-            "next_tmux_session_name": state.tmux_session_name.as_deref(),
-            "root": root.display().to_string(),
-            "path": path.display().to_string(),
-        }),
-    );
-
-    // #3416 (#3089 B3): observe→ENFORCE under the durable-authority flag (no-op
-    // when OFF); see dr::authority_blocks_backward_inflight_write. The violation
-    // itself was already recorded by the monotonic record_inflight_invariant
-    // above (downgraded to WARN for this skipped-write case — see #3552).
-    if enforce_skips_backward_write {
-        tracing::warn!(
-            "#3416 enforce: skipped backward inflight write at {}",
-            path.display()
-        );
-        return false;
-    }
-    true
-}
-
-pub(super) fn save_inflight_state(state: &InflightTurnState) -> Result<(), String> {
-    let Some(root) = inflight_runtime_root() else {
-        return Err("Home directory not found".to_string());
-    };
-    save_inflight_state_in_root(&root, state)
-}
-
-/// #897 counter-model review P2 #1 — atomic "create, don't overwrite"
-/// variant of `save_inflight_state`. Used by `POST /api/inflight/rebind` so a
-/// concurrent legitimate turn that wins the mailbox race between the rebind
-/// handler's existence check and its write cannot have its canonical
-/// inflight file silently overwritten by the synthetic rebind state
-/// (`user_msg_id=0`, placeholder ids zeroed). Returns `InflightAlreadyExists`
-/// when the target path is already occupied — the handler translates that
-/// into HTTP 409 and the operator retries (or leaves it to the live turn).
-#[derive(Debug)]
-pub(super) enum CreateNewInflightError {
-    /// A state file already exists at the target path — another path wrote
-    /// it between the caller's preflight check and this call.
-    AlreadyExists,
-    /// Filesystem or serialization failure.
-    Internal(String),
-}
-
-impl std::fmt::Display for CreateNewInflightError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::AlreadyExists => write!(f, "inflight state already exists"),
-            Self::Internal(msg) => write!(f, "{msg}"),
-        }
-    }
-}
-
-pub(super) fn save_inflight_state_create_new(
-    state: &InflightTurnState,
-) -> Result<(), CreateNewInflightError> {
-    let Some(root) = inflight_runtime_root() else {
-        return Err(CreateNewInflightError::Internal(
-            "Home directory not found".to_string(),
-        ));
-    };
-    save_inflight_state_create_new_in_root(&root, state)
-}
-
-/// Test-visible inner form of `save_inflight_state_create_new`. Takes an
-/// explicit root so unit tests can exercise the O_CREAT|O_EXCL semantics
-/// without tripping over `AGENTDESK_ROOT_DIR` env-var races.
-fn save_inflight_state_create_new_in_root(
-    root: &Path,
-    state: &InflightTurnState,
-) -> Result<(), CreateNewInflightError> {
-    let Some(provider) = state.provider_kind() else {
-        return Err(CreateNewInflightError::Internal(format!(
-            "Unknown provider '{}'",
-            state.provider
-        )));
-    };
-    let path = inflight_state_path(root, &provider, state.channel_id);
-    if let Some(parent) = path.parent() {
-        fs::create_dir_all(parent).map_err(|e| CreateNewInflightError::Internal(e.to_string()))?;
-    }
-    let _lock = lock_inflight_state_path(&path).map_err(CreateNewInflightError::Internal)?;
-    let mut updated = state.clone();
-    updated.ensure_finalizer_turn_id();
-    validate_inflight_state_for_save(
-        root,
-        &path,
-        &updated,
-        "src/services/discord/inflight.rs:save_inflight_state_create_new_in_root",
-    );
-    updated.updated_at = now_string();
-    let json = serde_json::to_string_pretty(&updated)
-        .map_err(|e| CreateNewInflightError::Internal(e.to_string()))?;
-
-    // `OpenOptions::create_new(true)` is the canonical atomic check-and-
-    // create primitive on POSIX (O_CREAT | O_EXCL). No reliance on a
-    // preceding `load_inflight_state` — the kernel itself serializes this.
-    use std::io::Write;
-    match fs::OpenOptions::new()
-        .write(true)
-        .create_new(true)
-        .open(&path)
-    {
-        Ok(mut file) => {
-            file.write_all(json.as_bytes())
-                .map_err(|e| CreateNewInflightError::Internal(e.to_string()))?;
-            file.sync_all()
-                .map_err(|e| CreateNewInflightError::Internal(e.to_string()))?;
-            Ok(())
-        }
-        Err(e) if e.kind() == std::io::ErrorKind::AlreadyExists => {
-            Err(CreateNewInflightError::AlreadyExists)
-        }
-        Err(e) => Err(CreateNewInflightError::Internal(e.to_string())),
-    }
-}
-
-fn save_inflight_state_in_root(root: &Path, state: &InflightTurnState) -> Result<(), String> {
-    let Some(provider) = state.provider_kind() else {
-        return Err(format!("Unknown provider '{}'", state.provider));
-    };
-    let path = inflight_state_path(root, &provider, state.channel_id);
-    if let Some(parent) = path.parent() {
-        fs::create_dir_all(parent).map_err(|e| e.to_string())?;
-    }
-    let _lock = lock_inflight_state_path(&path)?;
-    let mut updated = state.clone();
-    updated.ensure_finalizer_turn_id();
-    if !validate_inflight_state_for_save(
-        root,
-        &path,
-        &updated,
-        "src/services/discord/inflight.rs:save_inflight_state_in_root",
-    ) {
-        return Ok(());
-    }
-    updated.updated_at = now_string();
-    let json = serde_json::to_string_pretty(&updated).map_err(|e| e.to_string())?;
-    atomic_write(&path, &json)
-}
-
-/// #3107 codex re-review (P1): atomic compare-and-set save. Writes `state` ONLY
-/// when no inflight row exists for `(provider, channel_id)`, returning `true` iff
-/// it wrote. The watcher self-heal re-acquire previously did a non-atomic
-/// `load(...).is_some()` preflight + unconditional save: a concurrent intake
-/// could create a REAL inflight in the gap, and the synthetic `user_msg_id = 0`
-/// save would clobber it (lost turn). This closes the window by doing the check
-/// AND write under the same `lock_inflight_state_path` flock the other save/clear
-/// paths serialize on, so the synthetic row is written only when there is
-/// genuinely no inflight at the moment of the atomic write.
-pub(super) fn save_inflight_state_if_absent(state: &InflightTurnState) -> Result<bool, String> {
-    let Some(root) = inflight_runtime_root() else {
-        return Err("Home directory not found".to_string());
-    };
-    save_inflight_state_if_absent_in_root(&root, state)
-}
-
-fn save_inflight_state_if_absent_in_root(
-    root: &Path,
-    state: &InflightTurnState,
-) -> Result<bool, String> {
-    let Some(provider) = state.provider_kind() else {
-        return Err(format!("Unknown provider '{}'", state.provider));
-    };
-    let path = inflight_state_path(root, &provider, state.channel_id);
-    if let Some(parent) = path.parent() {
-        fs::create_dir_all(parent).map_err(|e| e.to_string())?;
-    }
-    // Hold the sidecar flock across the existence check AND the write so a
-    // concurrent intake `save_inflight_state_in_root` (which takes the same
-    // lock) cannot land a real inflight in the gap. `path.exists()` under the
-    // lock is the compare; `atomic_write` is the set.
-    let _lock = lock_inflight_state_path(&path)?;
-    if path.exists() {
-        return Ok(false);
-    }
-    let mut updated = state.clone();
-    updated.ensure_finalizer_turn_id();
-    validate_inflight_state_for_save(
-        root,
-        &path,
-        &updated,
-        "src/services/discord/inflight.rs:save_inflight_state_if_absent_in_root",
-    );
-    updated.updated_at = now_string();
-    let json = serde_json::to_string_pretty(&updated).map_err(|e| e.to_string())?;
-    atomic_write(&path, &json)?;
-    Ok(true)
-}
-
-/// Outcome of [`save_inflight_state_if_matches_identity`] — the #3041 P1-2 R3
-/// identity-guarded re-save used on a delivery-lease `Skip` epilogue.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(in crate::services::discord) enum GuardedSaveOutcome {
-    /// On-disk row still matched the turn identity; the row was rewritten.
-    Saved,
-    /// No inflight row existed (the lease HOLDER already cleared it on its
-    /// success path). We do NOT resurrect it — the turn is already delivered.
-    Missing,
-    /// A row existed but its identity did NOT match (a newer turn replaced it,
-    /// or a planned-restart / rebind-origin marker now owns the row). We do
-    /// NOT clobber it.
-    IdentityMismatch,
-    /// Filesystem / serialization error during the write.
-    IoError,
-}
-
-/// #3041 P1-2 (codex P1-2 R3): identity-guarded re-save for the bridge's
-/// delivery-lease `Skip` epilogue. On a Skip the live HOLDER (watcher) owns the
-/// turn and CLEARS the row on success, so the bridge epilogue must NOT blindly
-/// `save_inflight_state`: if the holder's clear won the race, a blind re-save
-/// would resurrect a STALE row for an already-delivered turn (recovery then sees
-/// it delivered, never clears, leaks the row). This closes the window the same
-/// way `clear_inflight_state_if_matches` (#2427 D-wire) does: under the lock,
-/// write only when the row is STILL present AND its `(user_msg_id, started_at,
-/// tmux_session_name)` identity (+ `turn_start_offset` when known) matches. Gone
-/// (`Missing`) or replaced by a newer turn / restart-rebind marker
-/// (`IdentityMismatch`) → no-op; holder FAILED + didn't clear → still present &
-/// matching → refresh (`Saved`). Same flock + atomic_write primitives as the
-/// rest of the module (Windows-safe).
-pub(in crate::services::discord) fn save_inflight_state_if_matches_identity(
-    state: &InflightTurnState,
-    expected: &InflightTurnIdentity,
-    expected_turn_start_offset: Option<u64>,
-) -> GuardedSaveOutcome {
-    let Some(root) = inflight_runtime_root() else {
-        return GuardedSaveOutcome::IoError;
-    };
-    save_inflight_state_if_matches_identity_in_root(
-        &root,
-        state,
-        expected,
-        expected_turn_start_offset,
-    )
-}
-
-pub(in crate::services::discord) fn save_existing_inflight_rebind_adoption_if_matches_identity(
-    state: &InflightTurnState,
-    expected: &InflightTurnIdentity,
-    expected_turn_start_offset: Option<u64>,
-) -> GuardedSaveOutcome {
-    let Some(root) = inflight_runtime_root() else {
-        return GuardedSaveOutcome::IoError;
-    };
-    save_existing_inflight_rebind_adoption_if_matches_identity_in_root(
-        &root,
-        state,
-        expected,
-        expected_turn_start_offset,
-    )
-}
-
-pub(in crate::services::discord) fn save_existing_inflight_rebind_adoption_with_offset_rebase_if_matches_identity(
-    state: &InflightTurnState,
-    expected: &InflightTurnIdentity,
-    expected_turn_start_offset: Option<u64>,
-    expected_last_offset: u64,
-) -> GuardedSaveOutcome {
-    let Some(root) = inflight_runtime_root() else {
-        return GuardedSaveOutcome::IoError;
-    };
-    save_existing_inflight_rebind_adoption_with_offset_rebase_if_matches_identity_in_root(
-        &root,
-        state,
-        expected,
-        expected_turn_start_offset,
-        expected_last_offset,
-    )
-}
-
-fn save_existing_inflight_rebind_adoption_if_matches_identity_in_root(
-    root: &Path,
-    state: &InflightTurnState,
-    expected: &InflightTurnIdentity,
-    expected_turn_start_offset: Option<u64>,
-) -> GuardedSaveOutcome {
-    save_existing_inflight_rebind_adoption_impl_in_root(
-        root,
-        state,
-        expected,
-        expected_turn_start_offset,
-        None,
-    )
-}
-
-fn save_existing_inflight_rebind_adoption_with_offset_rebase_if_matches_identity_in_root(
-    root: &Path,
-    state: &InflightTurnState,
-    expected: &InflightTurnIdentity,
-    expected_turn_start_offset: Option<u64>,
-    expected_last_offset: u64,
-) -> GuardedSaveOutcome {
-    save_existing_inflight_rebind_adoption_impl_in_root(
-        root,
-        state,
-        expected,
-        expected_turn_start_offset,
-        Some(expected_last_offset),
-    )
-}
-
-fn save_existing_inflight_rebind_adoption_impl_in_root(
-    root: &Path,
-    state: &InflightTurnState,
-    expected: &InflightTurnIdentity,
-    expected_turn_start_offset: Option<u64>,
-    expected_last_offset_for_rebase: Option<u64>,
-) -> GuardedSaveOutcome {
-    let Some(provider) = state.provider_kind() else {
-        return GuardedSaveOutcome::IoError;
-    };
-    let path = inflight_state_path(root, &provider, state.channel_id);
-    if let Some(parent) = path.parent() {
-        if fs::create_dir_all(parent).is_err() {
-            return GuardedSaveOutcome::IoError;
-        }
-    }
-    let Ok(_lock) = lock_inflight_state_path(&path) else {
-        return GuardedSaveOutcome::IoError;
-    };
-    let Ok(data) = fs::read_to_string(&path) else {
-        return GuardedSaveOutcome::Missing;
-    };
-    let Ok(on_disk) = serde_json::from_str::<InflightTurnState>(&data) else {
-        return GuardedSaveOutcome::IdentityMismatch;
-    };
-    if on_disk.rebind_origin {
-        return GuardedSaveOutcome::IdentityMismatch;
-    }
-    if on_disk.restart_mode != state.restart_mode {
-        return GuardedSaveOutcome::IdentityMismatch;
-    }
-    if expected.user_msg_id == 0 || !expected.matches_state(&on_disk) {
-        return GuardedSaveOutcome::IdentityMismatch;
-    }
-    if let Some(expected_offset) = expected_turn_start_offset {
-        if on_disk.turn_start_offset != Some(expected_offset) {
-            return GuardedSaveOutcome::IdentityMismatch;
-        }
-    }
-    if expected_last_offset_for_rebase
-        .is_some_and(|expected_last| on_disk.last_offset != expected_last)
-    {
-        return GuardedSaveOutcome::IdentityMismatch;
-    }
-
-    let mut updated = on_disk;
-    updated.tmux_session_name = state.tmux_session_name.clone();
-    updated.output_path = state.output_path.clone();
-    updated.input_fifo_path = state.input_fifo_path.clone();
-    updated.set_relay_owner_kind(state.effective_relay_owner_kind());
-    if expected_last_offset_for_rebase.is_some() {
-        updated.last_offset = state.last_offset;
-        updated.turn_start_offset = state.turn_start_offset;
-        updated.last_watcher_relayed_offset = state.last_watcher_relayed_offset;
-        updated.last_watcher_relayed_generation_mtime_ns =
-            state.last_watcher_relayed_generation_mtime_ns;
-    }
-    updated.ensure_finalizer_turn_id();
-    let _ = validate_inflight_state_for_save(
-        root,
-        &path,
-        &updated,
-        "src/services/discord/inflight.rs:save_existing_inflight_rebind_adoption_if_matches_identity_in_root",
-    );
-    updated.updated_at = now_string();
-    let Ok(json) = serde_json::to_string_pretty(&updated) else {
-        return GuardedSaveOutcome::IoError;
-    };
-    match atomic_write(&path, &json) {
-        Ok(()) => GuardedSaveOutcome::Saved,
-        Err(error) => {
-            tracing::warn!(
-                provider = %provider.as_str(),
-                channel = state.channel_id,
-                expected_user_msg_id = expected.user_msg_id,
-                error = %error,
-                "existing inflight rebind adoption save failed; leaving on-disk row untouched"
-            );
-            GuardedSaveOutcome::IoError
-        }
-    }
-}
-
-/// Root-explicit inner form of [`save_inflight_state_if_matches_identity`] for
-/// unit tests (avoids `AGENTDESK_ROOT_DIR` env-var races).
-pub(super) fn save_inflight_state_if_matches_identity_in_root(
-    root: &Path,
-    state: &InflightTurnState,
-    expected: &InflightTurnIdentity,
-    expected_turn_start_offset: Option<u64>,
-) -> GuardedSaveOutcome {
-    let Some(provider) = state.provider_kind() else {
-        return GuardedSaveOutcome::IoError;
-    };
-    let path = inflight_state_path(root, &provider, state.channel_id);
-    if let Some(parent) = path.parent() {
-        if fs::create_dir_all(parent).is_err() {
-            return GuardedSaveOutcome::IoError;
-        }
-    }
-    // Hold the sidecar flock across the read AND the write so a concurrent
-    // holder `clear_inflight_state` (which takes the same lock) cannot land its
-    // remove between our identity check and our write.
-    let Ok(_lock) = lock_inflight_state_path(&path) else {
-        return GuardedSaveOutcome::IoError;
-    };
-    // Holder already cleared the row on its success path → do NOT resurrect.
-    let Ok(data) = fs::read_to_string(&path) else {
-        return GuardedSaveOutcome::Missing;
-    };
-    let Ok(on_disk) = serde_json::from_str::<InflightTurnState>(&data) else {
-        // Malformed row: treat like a mismatch and do not clobber — the loader
-        // eviction path GCs malformed payloads on the next read.
-        return GuardedSaveOutcome::IdentityMismatch;
-    };
-    // A newer turn (different identity) or a planned-restart / rebind-origin
-    // marker now owns the row — never overwrite it with this preserved turn.
-    if on_disk.restart_mode.is_some() || on_disk.rebind_origin {
-        return GuardedSaveOutcome::IdentityMismatch;
-    }
-    if expected.user_msg_id == 0 || !expected.matches_state(&on_disk) {
-        return GuardedSaveOutcome::IdentityMismatch;
-    }
-    if let Some(expected_offset) = expected_turn_start_offset {
-        if on_disk.turn_start_offset != Some(expected_offset) {
-            return GuardedSaveOutcome::IdentityMismatch;
-        }
-    }
-    // #3089 B3: verdict observe-only here — this path already identity/offset-
-    // gates above; the #3416 backward vector is the plain overwrite tails.
-    let mut updated = state.clone();
-    updated.ensure_finalizer_turn_id();
-    let _ = validate_inflight_state_for_save(
-        root,
-        &path,
-        &updated,
-        "src/services/discord/inflight.rs:save_inflight_state_if_matches_identity_in_root",
-    );
-    updated.updated_at = now_string();
-    let Ok(json) = serde_json::to_string_pretty(&updated) else {
-        return GuardedSaveOutcome::IoError;
-    };
-    match atomic_write(&path, &json) {
-        Ok(()) => GuardedSaveOutcome::Saved,
-        Err(error) => {
-            tracing::warn!(
-                provider = %provider.as_str(),
-                channel = state.channel_id,
-                expected_user_msg_id = expected.user_msg_id,
-                error = %error,
-                "inflight identity-guarded save failed; leaving on-disk row untouched"
-            );
-            GuardedSaveOutcome::IoError
-        }
-    }
-}
-
-/// Reads + deserializes the inflight row at `path` while the caller holds the
-/// sidecar flock. Returns `None` on a missing/malformed file (same lenient
-/// posture as `load_inflight_state`).
-fn load_inflight_state_unlocked(path: &Path) -> Option<InflightTurnState> {
-    let data = fs::read_to_string(path).ok()?;
-    parse_inflight_state_content(&data).ok()
-}
-
-/// Shared lock-held persist tail: validate, optionally stamp `updated_at`,
-/// atomic-write. Caller must already hold `lock_inflight_state_path`.
-///
-/// `bump_updated_at` controls whether `updated_at` is reset to now. Real
-/// lifecycle mutations bump it (quiescence clock resets); an owner *correction*
-/// of a proven-dead orphan (#3982) preserves the old, already-stale timestamp so
-/// downstream ownerless-stale filters drop the row immediately on the next read
-/// instead of after another 300 s window.
-fn persist_under_lock_inner(
-    root: &Path,
-    path: &Path,
-    state: &InflightTurnState,
-    caller: &'static str,
-    bump_updated_at: bool,
-) -> Result<(), String> {
-    let mut updated = state.clone();
-    updated.ensure_finalizer_turn_id();
-    if !validate_inflight_state_for_save(root, path, &updated, caller) {
-        return Ok(());
-    }
-    if bump_updated_at {
-        updated.updated_at = now_string();
-    }
-    let json = serde_json::to_string_pretty(&updated).map_err(|e| e.to_string())?;
-    atomic_write(path, &json)
-}
-
-/// Shared lock-held persist tail: validate, stamp `updated_at`, atomic-write.
-/// Caller must already hold `lock_inflight_state_path`.
-fn persist_under_lock(
-    root: &Path,
-    path: &Path,
-    state: &InflightTurnState,
-    caller: &'static str,
-) -> Result<(), String> {
-    persist_under_lock_inner(root, path, state, caller, true)
-}
-
-/// Like [`persist_under_lock`] but preserves the row's existing `updated_at`
-/// instead of bumping it to now. Used by the #3982 orphan downgrade: the owner
-/// correction of a confirmed-dead orphan is not new lifecycle activity, so its
-/// quiescence clock must not be reset, or the triggering TUI-direct turn's fresh
-/// re-read would see a "fresh" row and keep aborting.
-fn persist_under_lock_preserving_updated_at(
-    root: &Path,
-    path: &Path,
-    state: &InflightTurnState,
-    caller: &'static str,
-) -> Result<(), String> {
-    persist_under_lock_inner(root, path, state, caller, false)
-}
-
-pub(crate) fn clear_inflight_state(provider: &ProviderKind, channel_id: u64) -> bool {
-    let Some(root) = inflight_runtime_root() else {
-        return false;
-    };
-    let path = inflight_state_path(&root, provider, channel_id);
-    let Ok(_lock) = lock_inflight_state_path(&path) else {
-        return false;
-    };
-    fs::remove_file(path).is_ok()
-}
-
-/// Outcome of an explicit-signal cleanup attempt that is guarded against
-/// racing the next turn's inflight write (#2427 Pitfall #1).
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(crate) enum GuardedClearOutcome {
-    /// File matched the expected `user_msg_id` and was removed.
-    Cleared,
-    /// File existed but a different `user_msg_id` was on disk — the next
-    /// turn already wrote its inflight, so we leave it alone.
-    UserMsgMismatch,
-    /// File on disk is a planned-restart marker (`restart_mode` set). The
-    /// caller is an explicit cleanup signal that fired for the previous
-    /// generation, so the marker must be preserved for recovery.
-    PlannedRestartSkipped,
-    /// File on disk is a rebind origin (`rebind_origin = true`). Its
-    /// lifetime is owned by `/api/inflight/rebind`, not the watcher /
-    /// turn-bridge, so the cleanup signal does not apply.
-    RebindOriginSkipped,
-    /// No inflight file existed (already cleared by a peer / never written).
-    Missing,
-    /// Filesystem error during the final `remove_file` step. Distinguished
-    /// from `Missing` so callers can surface the cleanup failure (warn/error
-    /// log + do NOT cancel the watcher, since the inflight is still on
-    /// disk and the next sweeper tick will retry). Codex review HIGH on
-    /// PR #2460: previously these errors were silently bucketed as Missing,
-    /// hiding broken cleanup from the operator while the 1800s safety-net
-    /// did the real work.
-    IoError,
-}
-
-/// Idempotent inflight cleanup driven by an *explicit* turn-completion
-/// signal (`TurnCompleted` emit, pane death detection, etc.). This is the
-/// #2427 D / A wire — the regular completion-path hook may have already
-/// cleared the file (Cleared turns into Missing), so we only act when the
-/// on-disk inflight still describes the turn we believe just finished.
-///
-/// Guards:
-/// * `expected_user_msg_id` — required to defeat the Pitfall #1 race. It
-///   matches either the Discord `user_msg_id` or the row's `finalizer_turn_id`;
-///   `0` is treated as "no guard available" and refused.
-/// * `restart_mode = Some(_)` — preserved (planned drain/hot-swap turns
-///   must survive across the dcserver restart they were saved for).
-/// * `rebind_origin = true` — preserved (Pitfall #5).
-pub(crate) fn clear_inflight_state_if_matches(
-    provider: &ProviderKind,
-    channel_id: u64,
-    expected_user_msg_id: u64,
-) -> GuardedClearOutcome {
-    let Some(root) = inflight_runtime_root() else {
-        return GuardedClearOutcome::Missing;
-    };
-    clear_inflight_state_if_matches_in_root(&root, provider, channel_id, expected_user_msg_id)
-}
-
-pub(in crate::services::discord) fn clear_inflight_state_if_matches_identity(
-    provider: &ProviderKind,
-    channel_id: u64,
-    expected: &InflightTurnIdentity,
-) -> GuardedClearOutcome {
-    let Some(root) = inflight_runtime_root() else {
-        return GuardedClearOutcome::Missing;
-    };
-    clear_inflight_state_if_matches_identity_in_root(&root, provider, channel_id, expected)
-}
-
-pub(in crate::services::discord) fn clear_rebind_origin_inflight_state_if_matches_identity(
-    provider: &ProviderKind,
-    channel_id: u64,
-    expected: &InflightTurnIdentity,
-) -> GuardedClearOutcome {
-    let Some(root) = inflight_runtime_root() else {
-        return GuardedClearOutcome::Missing;
-    };
-    clear_rebind_origin_inflight_state_if_matches_identity_in_root(
-        &root, provider, channel_id, expected,
-    )
-}
-
-pub(in crate::services::discord) fn clear_inflight_state_if_matches_identity_after_delivery(
-    provider: &ProviderKind,
-    channel_id: u64,
-    expected: &InflightTurnIdentity,
-    expected_turn_start_offset: Option<u64>,
-    full_response: &str,
-    response_sent_offset: usize,
-    last_offset: u64,
-) -> (GuardedClearOutcome, bool) {
-    let Some(root) = inflight_runtime_root() else {
-        return (GuardedClearOutcome::Missing, false);
-    };
-    clear_inflight_state_if_matches_identity_after_delivery_in_root(
-        &root,
-        provider,
-        channel_id,
-        expected,
-        expected_turn_start_offset,
-        full_response,
-        response_sent_offset,
-        last_offset,
-    )
-}
-
-#[allow(clippy::too_many_arguments)]
-pub(in crate::services::discord) fn refresh_inflight_last_offset_if_matches_identity(
-    provider: &ProviderKind,
-    channel_id: u64,
-    expected: &InflightTurnIdentity,
-    expected_turn_start_offset: Option<u64>,
-    output_path: &str,
-    expected_current_msg_id: Option<u64>,
-    last_offset: u64,
-    caller_owner: RelayOwnerKind,
-) -> bool {
-    let Some(root) = inflight_runtime_root() else {
-        return false;
-    };
-    refresh_inflight_last_offset_if_matches_identity_in_root(
-        &root,
-        provider,
-        channel_id,
-        expected,
-        expected_turn_start_offset,
-        output_path,
-        expected_current_msg_id,
-        last_offset,
-        caller_owner,
-    )
-}
-
-/// Root-explicit variant for unit tests. Production callers should use
-/// [`clear_inflight_state_if_matches`].
-pub(super) fn clear_inflight_state_if_matches_in_root(
-    root: &std::path::Path,
-    provider: &ProviderKind,
-    channel_id: u64,
-    expected_user_msg_id: u64,
-) -> GuardedClearOutcome {
-    let path = inflight_state_path(root, provider, channel_id);
-    let Ok(_lock) = lock_inflight_state_path(&path) else {
-        return GuardedClearOutcome::IoError;
-    };
-    let Ok(data) = fs::read_to_string(&path) else {
-        return GuardedClearOutcome::Missing;
-    };
-    let Ok(state) = parse_inflight_state_content(&data) else {
-        // Malformed file: treat like Missing — the loader-side eviction
-        // will GC the malformed payload on the next read.
-        return GuardedClearOutcome::Missing;
-    };
-    if state.restart_mode.is_some() {
-        return GuardedClearOutcome::PlannedRestartSkipped;
-    }
-    if state.rebind_origin {
-        return GuardedClearOutcome::RebindOriginSkipped;
-    }
-    if !state.matches_finalizer_turn_id(expected_user_msg_id) {
-        return GuardedClearOutcome::UserMsgMismatch;
-    }
-    // #2450: save and guarded-clear share the same sidecar lock, so the
-    // read/validate/unlink sequence below cannot race a concurrent
-    // atomic-write rename for a fresh turn.
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::MetadataExt;
-        let Ok(pre) = fs::metadata(&path) else {
-            return GuardedClearOutcome::Missing;
-        };
-        let Ok(post) = fs::metadata(&path) else {
-            return GuardedClearOutcome::Missing;
-        };
-        if pre.dev() != post.dev() || pre.ino() != post.ino() {
-            return GuardedClearOutcome::UserMsgMismatch;
-        }
-        // Final re-read + re-validate before unlink keeps the older
-        // corruption/mismatch protections intact while the sidecar lock
-        // closes the save-vs-clear race.
-        let Ok(reread) = fs::read_to_string(&path) else {
-            return GuardedClearOutcome::Missing;
-        };
-        let Ok(restate) = parse_inflight_state_content(&reread) else {
-            return GuardedClearOutcome::Missing;
-        };
-        if !restate.matches_finalizer_turn_id(expected_user_msg_id)
-            || restate.restart_mode.is_some()
-            || restate.rebind_origin
-        {
-            return GuardedClearOutcome::UserMsgMismatch;
-        }
-    }
-    match fs::remove_file(&path) {
-        Ok(()) => GuardedClearOutcome::Cleared,
-        Err(error) if error.kind() == std::io::ErrorKind::NotFound => GuardedClearOutcome::Missing,
-        Err(error) => {
-            tracing::warn!(
-                provider = %provider.as_str(),
-                channel = channel_id,
-                expected_user_msg_id = expected_user_msg_id,
-                error = %error,
-                "inflight guarded-clear remove_file failed; treating as IoError so sweeper retries"
-            );
-            GuardedClearOutcome::IoError
-        }
-    }
-}
-
-/// #3161 (codex P1): zero-id epilogue/guard cleanup that is STILL identity-safe.
-///
-/// A zero-id turn (recovery / external-input / cluster-relay synthesized;
-/// zero-normalized at [`optional_message_id`]) cannot be authenticated against a
-/// non-zero `expected_user_msg_id`, so [`clear_inflight_state_if_matches`]
-/// deliberately refuses (`expected_user_msg_id == 0` → `UserMsgMismatch`) to
-/// avoid blind-deleting a row it cannot prove ownership of. But a zero-id turn
-/// still legitimately owns *its own* row (whose on-disk `user_msg_id` is also 0)
-/// and must clean it up — recovery cleanup depends on this.
-///
-/// This helper closes that gap: it clears ONLY when the on-disk row's
-/// `user_msg_id` is itself 0 (a genuine zero-id-owned row). If a NEWER real
-/// (non-zero) identity turn has since written its row, the on-disk
-/// `user_msg_id != 0` and we return `UserMsgMismatch` — preserving the newer
-/// owner so its status panel can still complete. Planned-restart markers and
-/// rebind origins are preserved exactly like the non-zero guarded clear.
-pub(crate) fn clear_inflight_state_if_matches_zero_owned(
-    provider: &ProviderKind,
-    channel_id: u64,
-) -> GuardedClearOutcome {
-    let Some(root) = inflight_runtime_root() else {
-        return GuardedClearOutcome::Missing;
-    };
-    clear_inflight_state_if_matches_zero_owned_in_root(&root, provider, channel_id)
-}
-
-/// Root-explicit variant of [`clear_inflight_state_if_matches_zero_owned`] for
-/// unit tests.
-pub(super) fn clear_inflight_state_if_matches_zero_owned_in_root(
-    root: &std::path::Path,
-    provider: &ProviderKind,
-    channel_id: u64,
-) -> GuardedClearOutcome {
-    let path = inflight_state_path(root, provider, channel_id);
-    let Ok(_lock) = lock_inflight_state_path(&path) else {
-        return GuardedClearOutcome::IoError;
-    };
-    let Ok(data) = fs::read_to_string(&path) else {
-        return GuardedClearOutcome::Missing;
-    };
-    let Ok(state) = serde_json::from_str::<InflightTurnState>(&data) else {
-        return GuardedClearOutcome::Missing;
-    };
-    if state.restart_mode.is_some() {
-        return GuardedClearOutcome::PlannedRestartSkipped;
-    }
-    if state.rebind_origin {
-        return GuardedClearOutcome::RebindOriginSkipped;
-    }
-    // The only thing a zero-id turn may clear is a zero-id-owned row. A newer
-    // non-zero owner has `user_msg_id != 0` → preserve it.
-    if state.user_msg_id != 0 {
-        return GuardedClearOutcome::UserMsgMismatch;
-    }
-    match fs::remove_file(&path) {
-        Ok(()) => GuardedClearOutcome::Cleared,
-        Err(error) if error.kind() == std::io::ErrorKind::NotFound => GuardedClearOutcome::Missing,
-        Err(error) => {
-            tracing::warn!(
-                provider = %provider.as_str(),
-                channel = channel_id,
-                error = %error,
-                "inflight zero-owned guarded-clear remove_file failed; treating as IoError so sweeper retries"
-            );
-            GuardedClearOutcome::IoError
-        }
-    }
-}
-
-/// #3859: true when the row anchors a finalizable "🔄 처리 중" placeholder — a real
-/// placeholder message id (not 0, and not the placeholderless shape where the
-/// anchor mirrors the user's own message id), that is still a PURE placeholder
-/// (no streamed assistant content) or an explicitly-active long-running card.
-/// Mirrors the placeholder sweeper's abandon-eligibility gate so partial-output
-/// failure rows keep their delivered text (no "중단됨" clobber) exactly as the
-/// pre-#3859 path did.
-fn row_has_finalizable_placeholder(state: &InflightTurnState) -> bool {
-    if state.current_msg_id == 0
-        || (state.user_msg_id != 0 && state.current_msg_id == state.user_msg_id)
-    {
-        return false;
-    }
-    state.long_running_placeholder_active
-        || (state.full_response.is_empty() && state.response_sent_offset == 0)
-}
-
-/// #3859: record a durable abandon-request for `state`'s placeholder so the
-/// async `placeholder_sweeper` drain finalizes it to "중단됨" — enqueued UNDER the
-/// sidecar lock and BEFORE the caller's unlink, so the request survives even if
-/// the process dies right after the delete.
-///
-/// Returns `true` when it is SAFE for the caller to delete the inflight row:
-/// either the row anchors no finalizable placeholder (nothing to strand) OR the
-/// abandon-request was DURABLY persisted. Returns `false` ONLY when a finalizable
-/// placeholder's record FAILED to persist (#3859 r5 — codex P1); the caller MUST
-/// then keep the row so a later sweeper pass retries and the placeholder is never
-/// stranded without a record. Invariant: never `(row deleted ∧ record absent)`.
-#[must_use]
-fn enqueue_abandon_request_for_row(
-    provider: &ProviderKind,
-    channel_id: u64,
-    token_hash: &str,
-    state: &InflightTurnState,
-) -> bool {
-    if !row_has_finalizable_placeholder(state) {
-        return true;
-    }
-    match super::abandon_request_store::enqueue(
-        provider,
-        token_hash,
-        channel_id,
-        super::abandon_request_store::AbandonRecord {
-            msg_id: state.current_msg_id,
-            started_at: state.started_at.clone(),
-            current_tool_line: state.current_tool_line.clone(),
-        },
-    ) {
-        Ok(()) => true,
-        Err(error) => {
-            tracing::warn!(
-                provider = %provider.as_str(),
-                channel = channel_id,
-                msg_id = state.current_msg_id,
-                error = %error,
-                "abandon-request enqueue failed; PRESERVING inflight row so the placeholder is not stranded (sweeper retries next pass)"
-            );
-            false
-        }
-    }
-}
-
-/// #3859: failure-path cleanup that drives a stranded placeholder to a TERMINAL
-/// "중단됨" card WITHOUT keeping the inflight row alive.
-///
-/// Identical ownership guards to [`clear_inflight_state_if_matches`]
-/// (planned-restart / rebind-origin preserved; `UserMsgMismatch` for a newer
-/// owner; `expected_user_msg_id == 0` refused) — so a restart/rebind/foreign row
-/// is never enqueued or deleted. When the guards pass and the row anchors a
-/// finalizable placeholder, a durable abandon-request is enqueued (so the
-/// placeholder sweeper finalizes the "🔄 처리 중" card to "중단됨" by message id),
-/// then the row is DELETED — freeing the channel immediately like the pre-#3859
-/// path. The abandon-request is decoupled from the inflight lifecycle, so a
-/// re-adopt (new row + new placeholder) never collides with it (#3859 r4).
-pub(crate) fn request_inflight_abandon_if_matches(
-    provider: &ProviderKind,
-    channel_id: u64,
-    expected_user_msg_id: u64,
-    token_hash: &str,
-) -> GuardedClearOutcome {
-    let Some(root) = inflight_runtime_root() else {
-        return GuardedClearOutcome::Missing;
-    };
-    request_inflight_abandon_if_matches_in_root(
-        &root,
-        provider,
-        channel_id,
-        expected_user_msg_id,
-        token_hash,
-    )
-}
-
-/// Root-explicit variant for unit tests (the inflight ops use `root`; the
-/// abandon-request store is env-rooted via `discord_abandon_requests_root`, so a
-/// test must also set `AGENTDESK_ROOT_DIR`).
-pub(super) fn request_inflight_abandon_if_matches_in_root(
-    root: &std::path::Path,
-    provider: &ProviderKind,
-    channel_id: u64,
-    expected_user_msg_id: u64,
-    token_hash: &str,
-) -> GuardedClearOutcome {
-    let path = inflight_state_path(root, provider, channel_id);
-    let Ok(_lock) = lock_inflight_state_path(&path) else {
-        return GuardedClearOutcome::IoError;
-    };
-    let Ok(data) = fs::read_to_string(&path) else {
-        return GuardedClearOutcome::Missing;
-    };
-    let Ok(state) = parse_inflight_state_content(&data) else {
-        return GuardedClearOutcome::Missing;
-    };
-    if state.restart_mode.is_some() {
-        return GuardedClearOutcome::PlannedRestartSkipped;
-    }
-    if state.rebind_origin {
-        return GuardedClearOutcome::RebindOriginSkipped;
-    }
-    if !state.matches_finalizer_turn_id(expected_user_msg_id) {
-        return GuardedClearOutcome::UserMsgMismatch;
-    }
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::MetadataExt;
-        let Ok(pre) = fs::metadata(&path) else {
-            return GuardedClearOutcome::Missing;
-        };
-        let Ok(post) = fs::metadata(&path) else {
-            return GuardedClearOutcome::Missing;
-        };
-        if pre.dev() != post.dev() || pre.ino() != post.ino() {
-            return GuardedClearOutcome::UserMsgMismatch;
-        }
-        let Ok(reread) = fs::read_to_string(&path) else {
-            return GuardedClearOutcome::Missing;
-        };
-        let Ok(restate) = parse_inflight_state_content(&reread) else {
-            return GuardedClearOutcome::Missing;
-        };
-        if !restate.matches_finalizer_turn_id(expected_user_msg_id)
-            || restate.restart_mode.is_some()
-            || restate.rebind_origin
-        {
-            return GuardedClearOutcome::UserMsgMismatch;
-        }
-    }
-    // Enqueue BEFORE unlink (durable handoff). #3859 r5 (codex P1): if a
-    // FINALIZABLE placeholder's record fails to persist, DO NOT delete the row —
-    // return IoError so the sweeper retries (the row stays alive and the
-    // placeholder is finalized later). Never delete the row without its record.
-    if !enqueue_abandon_request_for_row(provider, channel_id, token_hash, &state) {
-        return GuardedClearOutcome::IoError;
-    }
-    match fs::remove_file(&path) {
-        Ok(()) => GuardedClearOutcome::Cleared,
-        Err(error) if error.kind() == std::io::ErrorKind::NotFound => GuardedClearOutcome::Missing,
-        Err(error) => {
-            tracing::warn!(
-                provider = %provider.as_str(),
-                channel = channel_id,
-                expected_user_msg_id = expected_user_msg_id,
-                error = %error,
-                "inflight abandon-request remove_file failed; treating as IoError so sweeper retries"
-            );
-            GuardedClearOutcome::IoError
-        }
-    }
-}
-
-/// #3859 zero-id variant — mirrors [`clear_inflight_state_if_matches_zero_owned`]
-/// guards, then enqueues an abandon-request (if the row anchors a finalizable
-/// placeholder) and deletes the row.
-pub(crate) fn request_inflight_abandon_if_matches_zero_owned(
-    provider: &ProviderKind,
-    channel_id: u64,
-    token_hash: &str,
-) -> GuardedClearOutcome {
-    let Some(root) = inflight_runtime_root() else {
-        return GuardedClearOutcome::Missing;
-    };
-    request_inflight_abandon_if_matches_zero_owned_in_root(&root, provider, channel_id, token_hash)
-}
-
-/// Root-explicit variant of [`request_inflight_abandon_if_matches_zero_owned`].
-pub(super) fn request_inflight_abandon_if_matches_zero_owned_in_root(
-    root: &std::path::Path,
-    provider: &ProviderKind,
-    channel_id: u64,
-    token_hash: &str,
-) -> GuardedClearOutcome {
-    let path = inflight_state_path(root, provider, channel_id);
-    let Ok(_lock) = lock_inflight_state_path(&path) else {
-        return GuardedClearOutcome::IoError;
-    };
-    let Ok(data) = fs::read_to_string(&path) else {
-        return GuardedClearOutcome::Missing;
-    };
-    let Ok(state) = serde_json::from_str::<InflightTurnState>(&data) else {
-        return GuardedClearOutcome::Missing;
-    };
-    if state.restart_mode.is_some() {
-        return GuardedClearOutcome::PlannedRestartSkipped;
-    }
-    if state.rebind_origin {
-        return GuardedClearOutcome::RebindOriginSkipped;
-    }
-    if state.user_msg_id != 0 {
-        return GuardedClearOutcome::UserMsgMismatch;
-    }
-    // #3859 r5: preserve the row if a finalizable placeholder's record fails to
-    // persist (never delete the row without its abandon-request).
-    if !enqueue_abandon_request_for_row(provider, channel_id, token_hash, &state) {
-        return GuardedClearOutcome::IoError;
-    }
-    match fs::remove_file(&path) {
-        Ok(()) => GuardedClearOutcome::Cleared,
-        Err(error) if error.kind() == std::io::ErrorKind::NotFound => GuardedClearOutcome::Missing,
-        Err(error) => {
-            tracing::warn!(
-                provider = %provider.as_str(),
-                channel = channel_id,
-                error = %error,
-                "inflight zero-owned abandon-request remove_file failed; treating as IoError so sweeper retries"
-            );
-            GuardedClearOutcome::IoError
-        }
-    }
-}
-
-fn clear_inflight_state_if_matches_identity_in_root(
-    root: &std::path::Path,
-    provider: &ProviderKind,
-    channel_id: u64,
-    expected: &InflightTurnIdentity,
-) -> GuardedClearOutcome {
-    let path = inflight_state_path(root, provider, channel_id);
-    let Ok(_lock) = lock_inflight_state_path(&path) else {
-        return GuardedClearOutcome::IoError;
-    };
-    let Ok(data) = fs::read_to_string(&path) else {
-        return GuardedClearOutcome::Missing;
-    };
-    let Ok(state) = serde_json::from_str::<InflightTurnState>(&data) else {
-        return GuardedClearOutcome::Missing;
-    };
-    if state.restart_mode.is_some() {
-        return GuardedClearOutcome::PlannedRestartSkipped;
-    }
-    if state.rebind_origin {
-        return GuardedClearOutcome::RebindOriginSkipped;
-    }
-    if expected.user_msg_id == 0 || !expected.matches_state(&state) {
-        return GuardedClearOutcome::UserMsgMismatch;
-    }
-    match fs::remove_file(&path) {
-        Ok(()) => GuardedClearOutcome::Cleared,
-        Err(error) if error.kind() == std::io::ErrorKind::NotFound => GuardedClearOutcome::Missing,
-        Err(error) => {
-            tracing::warn!(
-                provider = %provider.as_str(),
-                channel = channel_id,
-                expected_user_msg_id = expected.user_msg_id,
-                error = %error,
-                "inflight identity-guarded clear remove_file failed; treating as IoError so sweeper retries"
-            );
-            GuardedClearOutcome::IoError
-        }
-    }
-}
-
-fn clear_rebind_origin_inflight_state_if_matches_identity_in_root(
-    root: &std::path::Path,
-    provider: &ProviderKind,
-    channel_id: u64,
-    expected: &InflightTurnIdentity,
-) -> GuardedClearOutcome {
-    let path = inflight_state_path(root, provider, channel_id);
-    let Ok(_lock) = lock_inflight_state_path(&path) else {
-        return GuardedClearOutcome::IoError;
-    };
-    let Ok(data) = fs::read_to_string(&path) else {
-        return GuardedClearOutcome::Missing;
-    };
-    let Ok(state) = serde_json::from_str::<InflightTurnState>(&data) else {
-        return GuardedClearOutcome::Missing;
-    };
-    if state.restart_mode.is_some() {
-        return GuardedClearOutcome::PlannedRestartSkipped;
-    }
-    if !state.rebind_origin {
-        return GuardedClearOutcome::UserMsgMismatch;
-    }
-    if !expected.matches_state(&state) {
-        return GuardedClearOutcome::UserMsgMismatch;
-    }
-    match fs::remove_file(&path) {
-        Ok(()) => GuardedClearOutcome::Cleared,
-        Err(error) if error.kind() == std::io::ErrorKind::NotFound => GuardedClearOutcome::Missing,
-        Err(error) => {
-            tracing::warn!(
-                provider = %provider.as_str(),
-                channel = channel_id,
-                expected_user_msg_id = expected.user_msg_id,
-                error = %error,
-                "rebind-origin inflight guarded-clear remove_file failed; treating as IoError so sweeper retries"
-            );
-            GuardedClearOutcome::IoError
-        }
-    }
-}
-
-fn normalize_response_sent_offset(full_response: &str, response_sent_offset: usize) -> usize {
-    let mut offset = response_sent_offset.min(full_response.len());
-    while offset > 0 && !full_response.is_char_boundary(offset) {
-        offset -= 1;
-    }
-    offset
-}
-
-fn clear_inflight_state_if_matches_identity_after_delivery_in_root(
-    root: &std::path::Path,
-    provider: &ProviderKind,
-    channel_id: u64,
-    expected: &InflightTurnIdentity,
-    expected_turn_start_offset: Option<u64>,
-    full_response: &str,
-    response_sent_offset: usize,
-    last_offset: u64,
-) -> (GuardedClearOutcome, bool) {
-    let path = inflight_state_path(root, provider, channel_id);
-    let Ok(_lock) = lock_inflight_state_path(&path) else {
-        return (GuardedClearOutcome::IoError, false);
-    };
-    let Ok(data) = fs::read_to_string(&path) else {
-        return (GuardedClearOutcome::Missing, false);
-    };
-    let Ok(state) = serde_json::from_str::<InflightTurnState>(&data) else {
-        return (GuardedClearOutcome::Missing, false);
-    };
-    if state.restart_mode.is_some() {
-        return (GuardedClearOutcome::PlannedRestartSkipped, false);
-    }
-    if state.rebind_origin {
-        return (GuardedClearOutcome::RebindOriginSkipped, false);
-    }
-    if expected.user_msg_id == 0 || !expected.matches_state(&state) {
-        return (GuardedClearOutcome::UserMsgMismatch, false);
-    }
-    if let Some(expected_offset) = expected_turn_start_offset {
-        if state.turn_start_offset != Some(expected_offset) {
-            return (GuardedClearOutcome::UserMsgMismatch, false);
-        }
-    }
-
-    let mut delivered_state = state;
-    delivered_state.full_response = full_response.to_string();
-    delivered_state.response_sent_offset =
-        normalize_response_sent_offset(full_response, response_sent_offset);
-    delivered_state.last_offset = last_offset;
-    delivered_state.ensure_finalizer_turn_id();
-    delivered_state.updated_at = now_string();
-
-    let mirrored_delivery = match serde_json::to_string_pretty(&delivered_state)
-        .map_err(|error| error.to_string())
-        .and_then(|json| atomic_write(&path, &json))
-    {
-        Ok(()) => true,
-        Err(error) => {
-            tracing::warn!(
-                provider = %provider.as_str(),
-                channel = channel_id,
-                expected_user_msg_id = expected.user_msg_id,
-                error = %error,
-                "inflight delivery mirror failed before identity-guarded clear"
-            );
-            false
-        }
-    };
-
-    let outcome = match fs::remove_file(&path) {
-        Ok(()) => GuardedClearOutcome::Cleared,
-        Err(error) if error.kind() == std::io::ErrorKind::NotFound => GuardedClearOutcome::Missing,
-        Err(error) => {
-            tracing::warn!(
-                provider = %provider.as_str(),
-                channel = channel_id,
-                expected_user_msg_id = expected.user_msg_id,
-                error = %error,
-                "inflight identity-guarded delivery clear remove_file failed; treating as IoError so sweeper retries"
-            );
-            GuardedClearOutcome::IoError
-        }
-    };
-    (outcome, mirrored_delivery)
-}
-
-// #3034: root-parameterized variant exercised directly by the unit tests
-// (the prod wrapper was removed; tests drive a tempdir root). Test-only seam.
-#[allow(dead_code)]
-fn clear_inflight_state_if_matches_tmux_response_in_root(
-    root: &std::path::Path,
-    provider: &ProviderKind,
-    channel_id: u64,
-    tmux_session_name: &str,
-    response: &str,
-) -> GuardedClearOutcome {
-    let tmux_session_name = tmux_session_name.trim();
-    let response = response.trim();
-    if tmux_session_name.is_empty() || response.is_empty() {
-        return GuardedClearOutcome::UserMsgMismatch;
-    }
-
-    let path = inflight_state_path(root, provider, channel_id);
-    let Ok(_lock) = lock_inflight_state_path(&path) else {
-        return GuardedClearOutcome::IoError;
-    };
-    let Ok(data) = fs::read_to_string(&path) else {
-        return GuardedClearOutcome::Missing;
-    };
-    let Ok(state) = serde_json::from_str::<InflightTurnState>(&data) else {
-        return GuardedClearOutcome::Missing;
-    };
-    if state.restart_mode.is_some() {
-        return GuardedClearOutcome::PlannedRestartSkipped;
-    }
-    if state.rebind_origin {
-        return GuardedClearOutcome::RebindOriginSkipped;
-    }
-    if state.tmux_session_name.as_deref().map(str::trim) != Some(tmux_session_name) {
-        return GuardedClearOutcome::UserMsgMismatch;
-    }
-    if state.full_response.trim() != response {
-        return GuardedClearOutcome::UserMsgMismatch;
-    }
-
-    match fs::remove_file(&path) {
-        Ok(()) => GuardedClearOutcome::Cleared,
-        Err(error) if error.kind() == std::io::ErrorKind::NotFound => GuardedClearOutcome::Missing,
-        Err(error) => {
-            tracing::warn!(
-                provider = %provider.as_str(),
-                channel = channel_id,
-                tmux_session_name,
-                error = %error,
-                "inflight tmux-response guarded clear remove_file failed; treating as IoError so sweeper retries"
-            );
-            GuardedClearOutcome::IoError
-        }
-    }
-}
-
-#[allow(clippy::too_many_arguments)]
-fn refresh_inflight_last_offset_if_matches_identity_in_root(
-    root: &std::path::Path,
-    provider: &ProviderKind,
-    channel_id: u64,
-    expected: &InflightTurnIdentity,
-    expected_turn_start_offset: Option<u64>,
-    output_path: &str,
-    expected_current_msg_id: Option<u64>,
-    last_offset: u64,
-    caller_owner: RelayOwnerKind,
-) -> bool {
-    let path = inflight_state_path(root, provider, channel_id);
-    let Ok(_lock) = lock_inflight_state_path(&path) else {
-        return false;
-    };
-    let Ok(data) = fs::read_to_string(&path) else {
-        return false;
-    };
-    let Ok(mut state) = serde_json::from_str::<InflightTurnState>(&data) else {
-        return false;
-    };
-    if state.restart_mode.is_some() || state.rebind_origin {
-        return false;
-    }
-    if state.output_path.as_deref() != Some(output_path) {
-        return false;
-    }
-    if let Some(expected_msg_id) = expected_current_msg_id {
-        if state.current_msg_id != expected_msg_id {
-            return false;
-        }
-    }
-    if expected.user_msg_id == 0 || !expected.matches_state(&state) {
-        return false;
-    }
-    if let Some(expected_offset) = expected_turn_start_offset {
-        if state.turn_start_offset != Some(expected_offset) {
-            return false;
-        }
-    }
-
-    // I6 (last_offset_owner_gated): the persisted watermark is advanced only
-    // by the current relay owner. A non-owner caller (standby/idle) follows
-    // the authoritative offset read-only and must yield to a live owner. The
-    // identity guards above already proved this is the SAME turn, so a live
-    // owner that differs from the caller is an authority conflict, not a
-    // fresh-turn reset.
-    let persisted_owner = state.effective_relay_owner_kind();
-    let owner_is_live = !matches!(persisted_owner, RelayOwnerKind::None);
-    if owner_is_live && persisted_owner != caller_owner {
-        record_inflight_invariant(
-            false,
-            &state,
-            "last_offset_owner_gated",
-            "src/services/discord/inflight.rs:refresh_inflight_last_offset_if_matches_identity_in_root",
-            "inflight last_offset must only be advanced by the current relay owner",
-            serde_json::json!({
-                "persisted_owner": persisted_owner.as_str(),
-                "caller_owner": caller_owner.as_str(),
-                "previous": state.last_offset,
-                "next": last_offset,
-                "path": path.display().to_string(),
-            }),
-        );
-        return false;
-    }
-
-    // I6 (last_offset_monotonic): same identity, so a backward watermark write
-    // would clobber the authoritative offset and replay a stale transcript
-    // tail. Reject and record. A fresh-turn reset is already excluded by the
-    // identity guards above.
-    if last_offset < state.last_offset {
-        record_inflight_invariant(
-            false,
-            &state,
-            "last_offset_monotonic",
-            "src/services/discord/inflight.rs:refresh_inflight_last_offset_if_matches_identity_in_root",
-            "inflight last_offset must not move backwards for the same turn identity",
-            serde_json::json!({
-                "previous": state.last_offset,
-                "next": last_offset,
-                "path": path.display().to_string(),
-            }),
-        );
-        return false;
-    }
-
-    state.last_offset = last_offset;
-    state.ensure_finalizer_turn_id();
-    state.updated_at = now_string();
-    serde_json::to_string_pretty(&state)
-        .map_err(|error| error.to_string())
-        .and_then(|json| atomic_write(&path, &json))
-        .is_ok()
 }
 
 fn inflight_state_allows_idle_tmux_repair_state(state: &InflightTurnState) -> bool {
