+++ src/services/discord/inflight/save_store.rs
@@ -1,0 +1,459 @@
+//! Inflight store-side CAS "save" cluster (#3835 extraction).
+//!
+//! The compare-and-set write half of the inflight sidecar contract: fresh-row
+//! creation, absent-guarded seeding, and the identity-guarded save / rebind-
+//! adoption variants. Moved verbatim out of `inflight.rs` so the hot state
+//! parent stays below the frozen production-LoC baseline without changing any
+//! call-site name. The parent re-exports every public symbol at its original
+//! visibility, so `inflight::*` flat paths stay byte-identical for discord-
+//! module / inflight-core callers. The `_in_root` explicit-root seams keep the
+//! `pub(super)` visibility the parent's tests (and sibling `budget` /
+//! `anchor_repost` test modules) reach via re-import. Offset-monotonic /
+//! identity-guard invariants are unchanged (pure move).
+
+use super::*;
+
+pub(in crate::services::discord) fn save_inflight_state(
+    state: &InflightTurnState,
+) -> Result<(), String> {
+    let Some(root) = inflight_runtime_root() else {
+        return Err("Home directory not found".to_string());
+    };
+    save_inflight_state_in_root(&root, state)
+}
+
+/// #897 counter-model review P2 #1 — atomic "create, don't overwrite"
+/// variant of `save_inflight_state`. Used by `POST /api/inflight/rebind` so a
+/// concurrent legitimate turn that wins the mailbox race between the rebind
+/// handler's existence check and its write cannot have its canonical
+/// inflight file silently overwritten by the synthetic rebind state
+/// (`user_msg_id=0`, placeholder ids zeroed). Returns `InflightAlreadyExists`
+/// when the target path is already occupied — the handler translates that
+/// into HTTP 409 and the operator retries (or leaves it to the live turn).
+#[derive(Debug)]
+pub(in crate::services::discord) enum CreateNewInflightError {
+    /// A state file already exists at the target path — another path wrote
+    /// it between the caller's preflight check and this call.
+    AlreadyExists,
+    /// Filesystem or serialization failure.
+    Internal(String),
+}
+
+impl std::fmt::Display for CreateNewInflightError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Self::AlreadyExists => write!(f, "inflight state already exists"),
+            Self::Internal(msg) => write!(f, "{msg}"),
+        }
+    }
+}
+
+pub(in crate::services::discord) fn save_inflight_state_create_new(
+    state: &InflightTurnState,
+) -> Result<(), CreateNewInflightError> {
+    let Some(root) = inflight_runtime_root() else {
+        return Err(CreateNewInflightError::Internal(
+            "Home directory not found".to_string(),
+        ));
+    };
+    save_inflight_state_create_new_in_root(&root, state)
+}
+
+/// Test-visible inner form of `save_inflight_state_create_new`. Takes an
+/// explicit root so unit tests can exercise the O_CREAT|O_EXCL semantics
+/// without tripping over `AGENTDESK_ROOT_DIR` env-var races.
+fn save_inflight_state_create_new_in_root(
+    root: &Path,
+    state: &InflightTurnState,
+) -> Result<(), CreateNewInflightError> {
+    let Some(provider) = state.provider_kind() else {
+        return Err(CreateNewInflightError::Internal(format!(
+            "Unknown provider '{}'",
+            state.provider
+        )));
+    };
+    let path = inflight_state_path(root, &provider, state.channel_id);
+    if let Some(parent) = path.parent() {
+        fs::create_dir_all(parent).map_err(|e| CreateNewInflightError::Internal(e.to_string()))?;
+    }
+    let _lock = lock_inflight_state_path(&path).map_err(CreateNewInflightError::Internal)?;
+    let mut updated = state.clone();
+    updated.ensure_finalizer_turn_id();
+    validate_inflight_state_for_save(
+        root,
+        &path,
+        &updated,
+        "src/services/discord/inflight.rs:save_inflight_state_create_new_in_root",
+    );
+    updated.updated_at = now_string();
+    let json = serde_json::to_string_pretty(&updated)
+        .map_err(|e| CreateNewInflightError::Internal(e.to_string()))?;
+
+    // `OpenOptions::create_new(true)` is the canonical atomic check-and-
+    // create primitive on POSIX (O_CREAT | O_EXCL). No reliance on a
+    // preceding `load_inflight_state` — the kernel itself serializes this.
+    use std::io::Write;
+    match fs::OpenOptions::new()
+        .write(true)
+        .create_new(true)
+        .open(&path)
+    {
+        Ok(mut file) => {
+            file.write_all(json.as_bytes())
+                .map_err(|e| CreateNewInflightError::Internal(e.to_string()))?;
+            file.sync_all()
+                .map_err(|e| CreateNewInflightError::Internal(e.to_string()))?;
+            Ok(())
+        }
+        Err(e) if e.kind() == std::io::ErrorKind::AlreadyExists => {
+            Err(CreateNewInflightError::AlreadyExists)
+        }
+        Err(e) => Err(CreateNewInflightError::Internal(e.to_string())),
+    }
+}
+
+pub(super) fn save_inflight_state_in_root(
+    root: &Path,
+    state: &InflightTurnState,
+) -> Result<(), String> {
+    let Some(provider) = state.provider_kind() else {
+        return Err(format!("Unknown provider '{}'", state.provider));
+    };
+    let path = inflight_state_path(root, &provider, state.channel_id);
+    if let Some(parent) = path.parent() {
+        fs::create_dir_all(parent).map_err(|e| e.to_string())?;
+    }
+    let _lock = lock_inflight_state_path(&path)?;
+    let mut updated = state.clone();
+    updated.ensure_finalizer_turn_id();
+    if !validate_inflight_state_for_save(
+        root,
+        &path,
+        &updated,
+        "src/services/discord/inflight.rs:save_inflight_state_in_root",
+    ) {
+        return Ok(());
+    }
+    updated.updated_at = now_string();
+    let json = serde_json::to_string_pretty(&updated).map_err(|e| e.to_string())?;
+    atomic_write(&path, &json)
+}
+
+/// #3107 codex re-review (P1): atomic compare-and-set save. Writes `state` ONLY
+/// when no inflight row exists for `(provider, channel_id)`, returning `true` iff
+/// it wrote. The watcher self-heal re-acquire previously did a non-atomic
+/// `load(...).is_some()` preflight + unconditional save: a concurrent intake
+/// could create a REAL inflight in the gap, and the synthetic `user_msg_id = 0`
+/// save would clobber it (lost turn). This closes the window by doing the check
+/// AND write under the same `lock_inflight_state_path` flock the other save/clear
+/// paths serialize on, so the synthetic row is written only when there is
+/// genuinely no inflight at the moment of the atomic write.
+pub(in crate::services::discord) fn save_inflight_state_if_absent(
+    state: &InflightTurnState,
+) -> Result<bool, String> {
+    let Some(root) = inflight_runtime_root() else {
+        return Err("Home directory not found".to_string());
+    };
+    save_inflight_state_if_absent_in_root(&root, state)
+}
+
+fn save_inflight_state_if_absent_in_root(
+    root: &Path,
+    state: &InflightTurnState,
+) -> Result<bool, String> {
+    let Some(provider) = state.provider_kind() else {
+        return Err(format!("Unknown provider '{}'", state.provider));
+    };
+    let path = inflight_state_path(root, &provider, state.channel_id);
+    if let Some(parent) = path.parent() {
+        fs::create_dir_all(parent).map_err(|e| e.to_string())?;
+    }
+    // Hold the sidecar flock across the existence check AND the write so a
+    // concurrent intake `save_inflight_state_in_root` (which takes the same
+    // lock) cannot land a real inflight in the gap. `path.exists()` under the
+    // lock is the compare; `atomic_write` is the set.
+    let _lock = lock_inflight_state_path(&path)?;
+    if path.exists() {
+        return Ok(false);
+    }
+    let mut updated = state.clone();
+    updated.ensure_finalizer_turn_id();
+    validate_inflight_state_for_save(
+        root,
+        &path,
+        &updated,
+        "src/services/discord/inflight.rs:save_inflight_state_if_absent_in_root",
+    );
+    updated.updated_at = now_string();
+    let json = serde_json::to_string_pretty(&updated).map_err(|e| e.to_string())?;
+    atomic_write(&path, &json)?;
+    Ok(true)
+}
+
+/// Outcome of [`save_inflight_state_if_matches_identity`] — the #3041 P1-2 R3
+/// identity-guarded re-save used on a delivery-lease `Skip` epilogue.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub(in crate::services::discord) enum GuardedSaveOutcome {
+    /// On-disk row still matched the turn identity; the row was rewritten.
+    Saved,
+    /// No inflight row existed (the lease HOLDER already cleared it on its
+    /// success path). We do NOT resurrect it — the turn is already delivered.
+    Missing,
+    /// A row existed but its identity did NOT match (a newer turn replaced it,
+    /// or a planned-restart / rebind-origin marker now owns the row). We do
+    /// NOT clobber it.
+    IdentityMismatch,
+    /// Filesystem / serialization error during the write.
+    IoError,
+}
+
+/// #3041 P1-2 (codex P1-2 R3): identity-guarded re-save for the bridge's
+/// delivery-lease `Skip` epilogue. On a Skip the live HOLDER (watcher) owns the
+/// turn and CLEARS the row on success, so the bridge epilogue must NOT blindly
+/// `save_inflight_state`: if the holder's clear won the race, a blind re-save
+/// would resurrect a STALE row for an already-delivered turn (recovery then sees
+/// it delivered, never clears, leaks the row). This closes the window the same
+/// way `clear_inflight_state_if_matches` (#2427 D-wire) does: under the lock,
+/// write only when the row is STILL present AND its `(user_msg_id, started_at,
+/// tmux_session_name)` identity (+ `turn_start_offset` when known) matches. Gone
+/// (`Missing`) or replaced by a newer turn / restart-rebind marker
+/// (`IdentityMismatch`) → no-op; holder FAILED + didn't clear → still present &
+/// matching → refresh (`Saved`). Same flock + atomic_write primitives as the
+/// rest of the module (Windows-safe).
+pub(in crate::services::discord) fn save_inflight_state_if_matches_identity(
+    state: &InflightTurnState,
+    expected: &InflightTurnIdentity,
+    expected_turn_start_offset: Option<u64>,
+) -> GuardedSaveOutcome {
+    let Some(root) = inflight_runtime_root() else {
+        return GuardedSaveOutcome::IoError;
+    };
+    save_inflight_state_if_matches_identity_in_root(
+        &root,
+        state,
+        expected,
+        expected_turn_start_offset,
+    )
+}
+
+pub(in crate::services::discord) fn save_existing_inflight_rebind_adoption_if_matches_identity(
+    state: &InflightTurnState,
+    expected: &InflightTurnIdentity,
+    expected_turn_start_offset: Option<u64>,
+) -> GuardedSaveOutcome {
+    let Some(root) = inflight_runtime_root() else {
+        return GuardedSaveOutcome::IoError;
+    };
+    save_existing_inflight_rebind_adoption_if_matches_identity_in_root(
+        &root,
+        state,
+        expected,
+        expected_turn_start_offset,
+    )
+}
+
+pub(in crate::services::discord) fn save_existing_inflight_rebind_adoption_with_offset_rebase_if_matches_identity(
+    state: &InflightTurnState,
+    expected: &InflightTurnIdentity,
+    expected_turn_start_offset: Option<u64>,
+    expected_last_offset: u64,
+) -> GuardedSaveOutcome {
+    let Some(root) = inflight_runtime_root() else {
+        return GuardedSaveOutcome::IoError;
+    };
+    save_existing_inflight_rebind_adoption_with_offset_rebase_if_matches_identity_in_root(
+        &root,
+        state,
+        expected,
+        expected_turn_start_offset,
+        expected_last_offset,
+    )
+}
+
+pub(super) fn save_existing_inflight_rebind_adoption_if_matches_identity_in_root(
+    root: &Path,
+    state: &InflightTurnState,
+    expected: &InflightTurnIdentity,
+    expected_turn_start_offset: Option<u64>,
+) -> GuardedSaveOutcome {
+    save_existing_inflight_rebind_adoption_impl_in_root(
+        root,
+        state,
+        expected,
+        expected_turn_start_offset,
+        None,
+    )
+}
+
+pub(super) fn save_existing_inflight_rebind_adoption_with_offset_rebase_if_matches_identity_in_root(
+    root: &Path,
+    state: &InflightTurnState,
+    expected: &InflightTurnIdentity,
+    expected_turn_start_offset: Option<u64>,
+    expected_last_offset: u64,
+) -> GuardedSaveOutcome {
+    save_existing_inflight_rebind_adoption_impl_in_root(
+        root,
+        state,
+        expected,
+        expected_turn_start_offset,
+        Some(expected_last_offset),
+    )
+}
+
+fn save_existing_inflight_rebind_adoption_impl_in_root(
+    root: &Path,
+    state: &InflightTurnState,
+    expected: &InflightTurnIdentity,
+    expected_turn_start_offset: Option<u64>,
+    expected_last_offset_for_rebase: Option<u64>,
+) -> GuardedSaveOutcome {
+    let Some(provider) = state.provider_kind() else {
+        return GuardedSaveOutcome::IoError;
+    };
+    let path = inflight_state_path(root, &provider, state.channel_id);
+    if let Some(parent) = path.parent() {
+        if fs::create_dir_all(parent).is_err() {
+            return GuardedSaveOutcome::IoError;
+        }
+    }
+    let Ok(_lock) = lock_inflight_state_path(&path) else {
+        return GuardedSaveOutcome::IoError;
+    };
+    let Ok(data) = fs::read_to_string(&path) else {
+        return GuardedSaveOutcome::Missing;
+    };
+    let Ok(on_disk) = serde_json::from_str::<InflightTurnState>(&data) else {
+        return GuardedSaveOutcome::IdentityMismatch;
+    };
+    if on_disk.rebind_origin {
+        return GuardedSaveOutcome::IdentityMismatch;
+    }
+    if on_disk.restart_mode != state.restart_mode {
+        return GuardedSaveOutcome::IdentityMismatch;
+    }
+    if expected.user_msg_id == 0 || !expected.matches_state(&on_disk) {
+        return GuardedSaveOutcome::IdentityMismatch;
+    }
+    if let Some(expected_offset) = expected_turn_start_offset {
+        if on_disk.turn_start_offset != Some(expected_offset) {
+            return GuardedSaveOutcome::IdentityMismatch;
+        }
+    }
+    if expected_last_offset_for_rebase
+        .is_some_and(|expected_last| on_disk.last_offset != expected_last)
+    {
+        return GuardedSaveOutcome::IdentityMismatch;
+    }
+
+    let mut updated = on_disk;
+    updated.tmux_session_name = state.tmux_session_name.clone();
+    updated.output_path = state.output_path.clone();
+    updated.input_fifo_path = state.input_fifo_path.clone();
+    updated.set_relay_owner_kind(state.effective_relay_owner_kind());
+    if expected_last_offset_for_rebase.is_some() {
+        updated.last_offset = state.last_offset;
+        updated.turn_start_offset = state.turn_start_offset;
+        updated.last_watcher_relayed_offset = state.last_watcher_relayed_offset;
+        updated.last_watcher_relayed_generation_mtime_ns =
+            state.last_watcher_relayed_generation_mtime_ns;
+    }
+    updated.ensure_finalizer_turn_id();
+    let _ = validate_inflight_state_for_save(
+        root,
+        &path,
+        &updated,
+        "src/services/discord/inflight.rs:save_existing_inflight_rebind_adoption_if_matches_identity_in_root",
+    );
+    updated.updated_at = now_string();
+    let Ok(json) = serde_json::to_string_pretty(&updated) else {
+        return GuardedSaveOutcome::IoError;
+    };
+    match atomic_write(&path, &json) {
+        Ok(()) => GuardedSaveOutcome::Saved,
+        Err(error) => {
+            tracing::warn!(
+                provider = %provider.as_str(),
+                channel = state.channel_id,
+                expected_user_msg_id = expected.user_msg_id,
+                error = %error,
+                "existing inflight rebind adoption save failed; leaving on-disk row untouched"
+            );
+            GuardedSaveOutcome::IoError
+        }
+    }
+}
+
+/// Root-explicit inner form of [`save_inflight_state_if_matches_identity`] for
+/// unit tests (avoids `AGENTDESK_ROOT_DIR` env-var races).
+pub(super) fn save_inflight_state_if_matches_identity_in_root(
+    root: &Path,
+    state: &InflightTurnState,
+    expected: &InflightTurnIdentity,
+    expected_turn_start_offset: Option<u64>,
+) -> GuardedSaveOutcome {
+    let Some(provider) = state.provider_kind() else {
+        return GuardedSaveOutcome::IoError;
+    };
+    let path = inflight_state_path(root, &provider, state.channel_id);
+    if let Some(parent) = path.parent() {
+        if fs::create_dir_all(parent).is_err() {
+            return GuardedSaveOutcome::IoError;
+        }
+    }
+    // Hold the sidecar flock across the read AND the write so a concurrent
+    // holder `clear_inflight_state` (which takes the same lock) cannot land its
+    // remove between our identity check and our write.
+    let Ok(_lock) = lock_inflight_state_path(&path) else {
+        return GuardedSaveOutcome::IoError;
+    };
+    // Holder already cleared the row on its success path → do NOT resurrect.
+    let Ok(data) = fs::read_to_string(&path) else {
+        return GuardedSaveOutcome::Missing;
+    };
+    let Ok(on_disk) = serde_json::from_str::<InflightTurnState>(&data) else {
+        // Malformed row: treat like a mismatch and do not clobber — the loader
+        // eviction path GCs malformed payloads on the next read.
+        return GuardedSaveOutcome::IdentityMismatch;
+    };
+    // A newer turn (different identity) or a planned-restart / rebind-origin
+    // marker now owns the row — never overwrite it with this preserved turn.
+    if on_disk.restart_mode.is_some() || on_disk.rebind_origin {
+        return GuardedSaveOutcome::IdentityMismatch;
+    }
+    if expected.user_msg_id == 0 || !expected.matches_state(&on_disk) {
+        return GuardedSaveOutcome::IdentityMismatch;
+    }
+    if let Some(expected_offset) = expected_turn_start_offset {
+        if on_disk.turn_start_offset != Some(expected_offset) {
+            return GuardedSaveOutcome::IdentityMismatch;
+        }
+    }
+    // #3089 B3: verdict observe-only here — this path already identity/offset-
+    // gates above; the #3416 backward vector is the plain overwrite tails.
+    let mut updated = state.clone();
+    updated.ensure_finalizer_turn_id();
+    let _ = validate_inflight_state_for_save(
+        root,
+        &path,
+        &updated,
+        "src/services/discord/inflight.rs:save_inflight_state_if_matches_identity_in_root",
+    );
+    updated.updated_at = now_string();
+    let Ok(json) = serde_json::to_string_pretty(&updated) else {
+        return GuardedSaveOutcome::IoError;
+    };
+    match atomic_write(&path, &json) {
+        Ok(()) => GuardedSaveOutcome::Saved,
+        Err(error) => {
+            tracing::warn!(
+                provider = %provider.as_str(),
+                channel = state.channel_id,
+                expected_user_msg_id = expected.user_msg_id,
+                error = %error,
+                "inflight identity-guarded save failed; leaving on-disk row untouched"
+            );
+            GuardedSaveOutcome::IoError
+        }
+    }
+}
